+++ nodeconductor/structure/serializers.py
@@ -901,8 +901,6 @@ class BaseServiceSerializer(six.with_metaclass(ServiceSerializerMetaclass,
     certifications = ServiceCertificationSerializer(many=True, read_only=True, source='settings.certifications')
     name = serializers.ReadOnlyField(source='settings.name')
 
-    validate_name_length = MaxLengthValidator(150)
-
     class Meta(object):
         model = NotImplemented
         fields = (
@@ -1016,12 +1014,10 @@ class BaseServiceSerializer(six.with_metaclass(ServiceSerializerMetaclass,
                 if attrs.get('customer') != settings.customer:
                     raise serializers.ValidationError('Customer must match settings customer.')
 
-        if self.context['request'].method in ('POST', 'PUT', 'PATCH'):
-            name = self.initial_data.get('name')
-            if name:
-                self._validate_name(name)
-
         if self.context['request'].method == 'POST':
+            name = self.initial_data.get('name')
+            if not name or not name.strip():
+                raise serializers.ValidationError({'name': 'Name cannot be empty'})
             # Make shallow copy to protect from mutations
             settings_fields = self.Meta.settings_fields[:]
             create_settings = any([attrs.get(f) for f in settings_fields])
@@ -1077,11 +1073,6 @@ class BaseServiceSerializer(six.with_metaclass(ServiceSerializerMetaclass,
     def _validate_settings(self, settings):
         pass
 
-    def _validate_name(self, name):
-        self.validate_name_length(name)
-        if not name.strip():
-            raise serializers.ValidationError({'name': 'Name cannot be empty'})
-
     def get_resources_count(self, service):
         return self.get_resources_count_map[service.pk]
 
@@ -1117,26 +1108,6 @@ class BaseServiceSerializer(six.with_metaclass(ServiceSerializerMetaclass,
         if project and not spl_model.objects.filter(project=project, service=service).exists():
             spl_model.objects.create(project=project, service=service)
         return service
-
-    def update(self, instance, attrs):
-        name = self.initial_data.get('name')
-
-        # user should not be able to update settings name if he is not an owner of settings customer and not a staff
-        if name and self._user_has_permissions_to_change_name():
-            instance.settings.name = name
-            instance.settings.save()
-
-        return super(BaseServiceSerializer, self).update(instance, attrs)
-
-    def _user_has_permissions_to_change_name(self):
-        return 'request' in self.context and (self._user_is_owner_of_customer_settings() or
-                                              self.context['request'].user.is_staff)
-
-    def _user_is_owner_of_customer_settings(self):
-        service = self.instance
-        customer = service.settings.customer
-        return (customer and
-                customer.has_user(self.context['request'].user, models.CustomerRole.OWNER))
 
 
 class BaseServiceProjectLinkSerializer(PermissionFieldFilteringMixin,