+++ CHANGELOG.rst
@@ -3,7 +3,7 @@ CHANGELOG
 
 Next Release
 ------------
-- Update service settings name update policy: allow to customer owner or staff only.
+- Disable service setting name update via service endpoint.
 
 Release 0.126.2
 ---------------