+++ CHANGELOG.rst
@@ -3,6 +3,10 @@ CHANGELOG
 
 Next Release
 ------------
+Nothing yet.
+
+Release 0.127.0
+---------------
 - Disable service setting name update via service endpoint.
 
 Release 0.126.2